Use more portable word types phase 1
The codebase is now transitioning to hardware which requires
targeting rv64. For variable holding data such as pointers
and sizes, we should use architecture defined word-lengths.

+++ bit_utils.c
@@ -18,7 +18,7 @@ uint32_t big2little_l(uint32_t big) {
     return ret;
 }
 
-uint32_t big2little_dl(uint64_t big) {
+uint64_t big2little_dl(uint64_t big) {
     uint64_t ret = 0;
     ret |= ((big & 0xFF00000000000000U) >> 56U) << 0U;
     ret |= ((big & 0x00FF000000000000U) >> 48U) << 8U;
